Improve Error handling when there is no agent present.
Previously, the connector would throw a stack trace and stop
working properly if the factory agent was gone and a user
attempted an action, this patch adds logging and catches
this type of exceptionm as well as returning a 500 server error
to the caller, to be handled as desired. Future rev will return
the error string as well, but Conductor does nothing with errors
at this point anyway.